+++ src/entities/project/index.ts
@@ -4,4 +4,4 @@ export * from './model/useGetProjects';
 export { default as ProjectCard } from './ui/ProjectCard';
 export { default as ProjectDetailModal } from './ui/ProjectDetailModal';
 export { default as ProjectRequestCard } from './ui/ProjectRequestCard';
-export { default as CheckRequestStatus } from './ui/ㅁㄴㅇㄹ';
+export { default as CheckRequestStatus } from './ui/CheckRequestStatus';
